Remove akimboautoswitch
The problem is, that some players (especially SMG users) got on purpose
akimbo with intention to not use it, only in order to others don't get it.
Such behaviour is directed "piggishly" against other players and
impoverish the game/gameplay.
Now, if people don't want to use some akimbo, they should just to not
get it, for example jump over it.

DOCREF akimboautoswitch