Add option to disable persistent spectating
Now persistent spectating of a player is disabled for spectatepersistent=0,
enabled only in non-arena modes if equals 1, always enabled if equals 2
(default). For disabled persistent spectating:
1. spectate mode returns automatically to the person perspective after
spawn of some player, if earlier it was switched automatically to the fly
mode when all players in the match died;
2. after death of spectated player it switches to the random alive player,
and not this with the lowest client number.

DOCREF spectatepersistent